requirements.txt
Flask==0.10.1
Flask-PyMongo==0.3.0
Flask-RESTful==0.2.12
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
aniso8601==0.82
gunicorn==19.0.0
itsdangerous==0.24
pymongo==2.8
pytz==2014.4
six==1.10.0
numpy==1.13.0
nltk==3.2.4
spacy==1.8.2
tensorflow==1.1.0
python-rake==1.0.7
gensim==2.1.0
html5lib==1.0b8
